+++ source/common/modules/markdown-editor/parser/highlight-parser.ts
@@ -19,9 +19,19 @@ const MarkDelimiter: DelimiterType = {
   mark: 'HighlightMark' // No specific syntax node (for now due to backwards compatibility reasons)
 }
 
-function isWhitespace (c: number): boolean {
+/**
+ * Determines if the provided char is allowed to surround highlighting marks
+ *
+ * @param   {number}   c  The char code
+ *
+ * @return  {boolean}     Whether it's allowed before/after a highlight mark
+ */
+function allowedSurroundingChar (c: number): boolean {
   const char = String.fromCharCode(c)
-  return /\s/.test(char)
+  // Regex matches whitespace (Z), non-word-characters (math, currentcy; S), and
+  // general punctuation characters. NOTE: Partially "stolen" and adapted from
+  // the Lezer Markdown parser.
+  return /\p{Z}|\p{S}|\p{P}/u.test(char)
 }
 
 export const highlightParser: InlineParser = {
@@ -38,19 +48,20 @@ export const highlightParser: InlineParser = {
     }
 
     // A highlight marker is considered opening if it is at the beginning of the
-    // line (bol) or is preceded by whitespace. Furthermore, it must not be
-    // followed by whitespace and not be at the end of the line (eol).
+    // line (bol) or is preceded by whitespace or any punctuation or generally
+    // non-word characters. Furthermore, it must not be followed by whitespace
+    // (or punctuation/certain chars) and not be at the end of the line (eol).
     // For a highlight marker to be considered closing, it needs the opposite
     // requirements. This is why we need both checks (because otherwise closing
     // tags would allow preceding whitespace which would prompt Pandoc not to
     // render them).
     const bol = pos === ctx.offset
-    const eol = pos === ctx.end
-    const spaceBefore = bol || (pos > ctx.offset && isWhitespace(ctx.char(pos - 1)))
-    const spaceAfter = eol || (pos + 1 < ctx.end && isWhitespace(ctx.char(pos + 2)))
+    const eol = pos + 2 === ctx.end
+    const validBefore = bol || (pos > ctx.offset && allowedSurroundingChar(ctx.char(pos - 1)))
+    const validAfter = eol || (pos + 1 < ctx.end && allowedSurroundingChar(ctx.char(pos + 2)))
 
-    const isOpening = spaceBefore && !spaceAfter
-    const isClosing = !spaceBefore // No check for spaceAfter to allow for, e.g., "Some ==test==." (with punctuation afterwards)
+    const isOpening = validBefore && !validAfter
+    const isClosing = !validBefore && validAfter
 
     return ctx.addDelimiter(MarkDelimiter, pos, pos + 2, isOpening, isClosing)
   }