+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Fix highlight markers not appearing while the emphasis renderer is on when the
   user edits a highlighted span
 * Add pascal syntax highlighting to the code block autocomplete
+* Improved highlight detection around punctuation marks and other non-word
+  characters
 
 ## Under the Hood